consolidate mmit code into PiecewiseFunction cpp and h files
