+++ R/RcppExports.R
@@ -1,14 +1,14 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-### Compute a data.frame with one row for each distribution
-### implemented in the C++ code, and columns distribution.str,
-### parameters, description.
+#' Compute a data.frame with one row for each distribution
+#' implemented in the C++ code, and columns distribution.str,
+#' parameters, description.
 get_distribution_info <- function() {
     .Call(`_binsegRcpp_get_distribution_info`)
 }
 
-### Low-level interface to binary segmentation algorithm.
+#' Low-level interface to binary segmentation algorithm.
 binseg_interface <- function(data_vec, weight_vec, max_segments, min_segment_length, distribution_str, container_str, is_validation_vec, position_vec) {
     .Call(`_binsegRcpp_binseg_interface`, data_vec, weight_vec, max_segments, min_segment_length, distribution_str, container_str, is_validation_vec, position_vec)
 }
