+++ src/PiecewiseFunction.cpp
@@ -16,11 +16,51 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include <cmath>
 #include <iostream>
 #include <iterator>
+#include "PiecewiseFunction.h"
 
-#include "double_utils.h"
-#include "piecewise_function.h"
-//#include <Rcpp.h>
+Coefficients Coefficients::operator+(Coefficients &other) {
+    return Coefficients(this->quadratic + other.quadratic, this->linear + other.linear, this->constant + other.constant);
+}
 
+void Coefficients::operator+=(Coefficients &other) {
+    this->quadratic += other.quadratic;
+    this->linear += other.linear;
+    this->constant += other.constant;
+}
+
+Coefficients Coefficients::operator-(Coefficients &other) {
+    return Coefficients(this->quadratic - other.quadratic, this->linear - other.linear, this->constant - other.constant);
+}
+
+void Coefficients::operator-=(Coefficients &other) {
+    this->quadratic -= other.quadratic;
+    this->linear -= other.linear;
+    this->constant -= other.constant;
+}
+
+Coefficients Coefficients::operator*(double scalar) {
+    return Coefficients(this->quadratic * scalar, this->linear * scalar, this->constant * scalar);
+}
+
+void Coefficients::operator*=(double scalar) {
+    this->quadratic *= scalar;
+    this->linear *= scalar;
+    this->constant *= scalar;
+}
+
+Coefficients Coefficients::operator/(double scalar) {
+    return Coefficients(this->quadratic / scalar, this->linear / scalar, this->constant / scalar);
+}
+
+void Coefficients::operator/=(double scalar) {
+    this->quadratic /= scalar;
+    this->linear /= scalar;
+    this->constant /= scalar;
+}
+
+bool Coefficients::operator==(Coefficients &other){
+    return this->quadratic == other.quadratic && this->linear == other.linear && this->constant == other.constant;
+}
 
 /*
  * Printing coefficients
