+++ src/binseg.h
@@ -19,10 +19,10 @@ class Distribution;
 // one parameter (mean) because the model is normal change in mean
 // with constant variance but there could be more parameters for other
 // models (e.g., normal change in mean and variance).
-class MeanVarLoss {
+class ParamsLoss {
 public:
-  MeanVarLoss(Distribution*);
-  MeanVarLoss() {
+  ParamsLoss(Distribution*);
+  ParamsLoss() {
     loss = INFINITY;
   }
   double loss;
@@ -33,7 +33,7 @@ public:
 class Split {
 public:
   int this_end;//index of last data point on the first/before segment.
-  MeanVarLoss before, after;
+  ParamsLoss before, after;
   double get_loss(void) const {
     return before.loss + after.loss;
   }
@@ -74,8 +74,8 @@ public:
   std::string description;
   param_names_type param_names_vec;
   virtual Split get_best_split(Set&,int,int,int,int) = 0;
-  virtual double loss_for_params(Set&,MeanVarLoss&,int,int) = 0;
-  virtual MeanVarLoss estimate_params(Set&,int,int) = 0;
+  virtual double loss_for_params(Set&,ParamsLoss&,int,int) = 0;
+  virtual ParamsLoss estimate_params(Set&,int,int) = 0;
 };
 
 typedef std::unordered_map<std::string, Distribution*> dist_map_type;
