--- src/function_types.h
@@ -1,24 +1,0 @@
-/*
-MMIT: Max Margin Interval Trees
-Copyright (C) 2017 Toby Dylan Hocking, Alexandre Drouin
-This program is free software: you can redistribute it and/or modify
-        it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-This program is distributed in the hope that it will be useful,
-        but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-*/
-#ifndef CORE_FUNCTION_TYPES_H_H
-#define CORE_FUNCTION_TYPES_H_H
-
-enum FunctionType{
-    linear_hinge,
-    squared_hinge
-};
-
-#endif //CORE_FUNCTION_TYPES_H_H
